requirements.txt
cssselect==0.9.2
defusedxml==0.4.1
Django==1.9.8
django-allauth==0.26.1
django-braces==1.9.0
django-environ==0.4.0
django-rest-auth==0.7.0
django-webpack-loader==0.3.0
djangorestframework==3.3.3
httplib2==0.9.2
lxml==3.6.1
nltk==3.2.1
oauth2==1.9.0.post1
oauthlib==1.0.3
PyJWT==1.4.1
python-social-auth==0.2.19
python3-openid==3.0.10
requests==2.10.0
requests-oauthlib==0.6.2
rest-social-auth==0.5.0
selenium==2.53.6
six==1.10.0
splinter==0.7.3
